fix: Multiple payment page and dashboard improvements
Payment Page Fixes:
- Auto-select POS account when switching to POS payment method
- Auto-select single bank account when switching to card transfer
- Center text in payment amount input box
- Update sale tax/discount amounts when processing payments with tax/discount

Dashboard Improvements:
- Add sale state filtering (OPEN, CLOSED, CANCELED, all)
- Superusers can see canceled sales, others can see OPEN and CLOSED
- State filter triggers API reload for better performance

Report Creation Fixes:
- Fix COGS calculation bug: now correctly calculates Sum(quantity × material_cost)
  instead of Sum(material_cost) × Sum(quantity)
- Fix type annotations: return Decimal instead of int for variance properties

+++ backend/apps/sale/models/daily_report_model.py
@@ -222,12 +222,11 @@ class DailyReport(models.Model):
         return self.closing_cash_counted - self.opening_float
 
     @property
-    def expected_cash_total(self) -> int:
+    def expected_cash_total(self) -> Decimal:
         """
         Expected cash from payments.
         Calculated from payment_methods where method=CASH.
         """
-
         return (
             self.payment_methods.select_related("daily_report")
             .get(payment_method=SalePayment.PaymentMethod.CASH)
@@ -251,7 +250,7 @@ class DailyReport(models.Model):
         )
 
     @property
-    def card_transfer_variance(self) -> int:
+    def card_transfer_variance(self) -> Decimal:
         return (
             self.payment_methods.select_related("daily_report")
             .get(payment_method=SalePayment.PaymentMethod.CARD_TRANSFER)
@@ -259,9 +258,9 @@ class DailyReport(models.Model):
         )
 
     @property
-    def cash_variance(self) -> int:
+    def cash_variance(self) -> Decimal:
         """Cash variance = (closing - opening) - expected cash."""
-        return self.net_cash_received - self.expected_cash_total
+        return Decimal(self.net_cash_received) - self.expected_cash_total
 
     @property
     def total_variance(self) -> Decimal:
